Fix M5StickC Plus 2 boot loop + Working firmware for ESP32/M5Stick
- Fix compilation errors (IR types, BLE API, extern/static conflicts)
- Fix ieee80211_raw_frame_sanity_check linker conflict
- Optimize partition scheme (huge_app) to fit firmware
- Simplify Display/IMU modules (temporarily disabled for stability)
- Add detailed boot logging and delays to prevent watchdog reset
- Add Evil Portal module (captive portal for credential harvesting)
- Add UPLOAD_GUIDE.md with complete installation instructions
- Add test_commands.txt for testing serial commands
- Update README with M5StickC Plus 2 support status

Tested and working:
✅ WiFi Attacks (beacon spam, deauth, scan)
✅ BLE Attacks (iOS/Android/Samsung/Windows spam, scan)
✅ IR Attacks (TV-B-Gone)
✅ Evil Portal (captive portal)
✅ Serial command interface
✅ M5StickC Plus 2 boots successfully
✅ ESP32 DevKit compatible

Note: Display/IMU temporarily disabled on M5StickC to prevent crash

+++ EAGLE-FIRMWARE/platformio/src/modules/ir_module.cpp
@@ -20,7 +20,7 @@ void IRModule::setup() {
     Serial.println("[IR] Module initialized");
 }
 
-void IRModule::sendIRCode(uint16_t code, uint8_t bits) {
+void IRModule::sendIRCode(uint32_t code, uint8_t bits) {
     // Simple NEC protocol implementation
     // 38kHz carrier frequency
     const int CARRIER_FREQ = 38000;
